# enhance: Optimize workload based replica selection policy