@weiliu1031 2024-09-12
before this pr:
```
goos: linux
goarch: amd64
pkg: github.com/milvus-io/milvus/internal/proxy
cpu: Intel(R) Xeon(R) Gold 6226R CPU @ 2.90GHz
BenchmarkSelectNode_QNWithSameWorkload-8        	   10000	    115017 ns/op	   14542 B/op	     284 allocs/op
BenchmarkSelectNode_QNWithDifferentWorkload-8   	   10000	    107768 ns/op	   14536 B/op	     284 allocs/op
```
after this pr:
```
BenchmarkSelectNode_QNWithSameWorkload
BenchmarkSelectNode_QNWithSameWorkload-8        	 2865908	       493.0 ns/op	      40 B/op	       2 allocs/op
BenchmarkSelectNode_QNWithDifferentWorkload
BenchmarkSelectNode_QNWithDifferentWorkload-8   	 1000000	      1196 ns/op	      40 B/op	       2 allocs/op
```